Prototipo de sistema para hacer análisis estadístico de datos/Sistema_analisis_estadistico/requirements.txt
asgiref==3.7.2
contourpy==1.2.0
crispy-bootstrap4==2023.1
cycler==0.12.1
Django==4.2.7
django-crispy-forms==2.1
fonttools==4.49.0
joblib==1.3.2
kiwisolver==1.4.5
lxml==5.1.0
matplotlib==3.8.3
numpy==1.26.1
packaging==24.0
pandas==2.1.3
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3.post1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sk==0.0.1
sqlparse==0.4.4
threadpoolctl==3.3.0
typing_extensions==4.10.0
tzdata==2023.3
